+++ spec/resource_profile_spec.rb
@@ -132,7 +132,8 @@ describe 'Profile test' do
                                                 "title": "CEO",
                                                 "seniority": "C_LEVEL"
                                             },
-                                            business_contact_email: "employee@company.com"
+                                            business_contact_email: "employee@company.com",
+                                            doing_business_as: "Test DBA"
                                         }
                                     ))
     
@@ -171,7 +172,8 @@ describe 'Profile test' do
                                 "title": "CEO",
                                 "seniority": "C_LEVEL"
                             },
-                            business_contact_email: "employee@company.com"
+                            business_contact_email: "employee@company.com",
+                            doing_business_as: "Test DBA"
                          })
         end
         it 'update profile' do
@@ -198,7 +200,8 @@ describe 'Profile test' do
                                                   "email": "xyz123@plivo.com",
                                                   "title": "ugigwc",
                                                   "seniority": "admin"
-                                              }
+                                              },
+                                              doing_business_as: "Updated DBA"
                                           }
                                       ))
       
@@ -227,7 +230,8 @@ describe 'Profile test' do
                                 "email": "xyz123@plivo.com",
                                 "title": "ugigwc",
                                 "seniority": "admin"
-                            }
+                            },
+                            doing_business_as: "Updated DBA"
                            })
           end
 end